Fix: Making update twice an exception occurs.

+++ LiveUpdate/Dashboard.Shell/CopyDirectoryExtension.cs
@@ -7,23 +7,32 @@ using System.Threading.Tasks;
 
 namespace Dashboard.Shell
 {
-    public static class CopyDirectoryExtension
-    {
-        public static DirectoryInfo CopyTo(this DirectoryInfo sourceDir, string destinationPath, bool overwrite = false)
-        {
-            var sourcePath = sourceDir.FullName;
+   public static class CopyDirectoryExtension
+   {
+      public static DirectoryInfo CopyTo(this DirectoryInfo sourceDir, string destinationPath, bool overwrite = false)
+      {
+         var sourcePath = sourceDir.FullName;
 
-            var destination = new DirectoryInfo(destinationPath);
-
-            destination.Create();
-
-            foreach (var sourceSubDirPath in Directory.EnumerateDirectories(sourcePath, "*", SearchOption.AllDirectories))
-                Directory.CreateDirectory(sourceSubDirPath.Replace(sourcePath, destinationPath));
-
-            foreach (var file in Directory.EnumerateFiles(sourcePath, "*", SearchOption.AllDirectories))
-                File.Copy(file, file.Replace(sourcePath, destinationPath), overwrite);
-
+         var destination = new DirectoryInfo(destinationPath);
+         if (destination.Exists)
+         {
             return destination;
-        }
-    }
+         }
+
+         destination.Create();
+
+         if (!sourceDir.Exists)
+         {
+            return destination;
+         }
+
+         foreach (var sourceSubDirPath in Directory.EnumerateDirectories(sourcePath, "*", SearchOption.AllDirectories))
+            Directory.CreateDirectory(sourceSubDirPath.Replace(sourcePath, destinationPath));
+
+         foreach (var file in Directory.EnumerateFiles(sourcePath, "*", SearchOption.AllDirectories))
+            File.Copy(file, file.Replace(sourcePath, destinationPath), overwrite);
+
+         return destination;
+      }
+   }
 }
